requirements.txt
keras==2.9.0
numpy==1.21.6
opencv-python==4.6.0.66
spotipy~=2.20.0
pyyaml~=6.0